Add File - Open Folder menu
This allows a folder to be selected. On clicking Open in the dialog
a workspace file will be created with the .fws file extension.
No files or folders are currently displayed in the Solution window.
The recently opened folder will show in the Recent Files list with
the text "[Folder]" displayed next to it.