Share code to add new file to workspace or folder
Added a command handler to the workspace node builder extension.
This new command handler, along with the existing workspace folder
command handler, now uses a common base class where the logic for
adding a new file is shared. Using a global command handler that
was not a node command handler for the workspace menu items resulted
in extra logic to find the solution window. This is no longer needed.